requirements.txt
python-dotenv
requests
faker
pycountry
